br 10
minor fixes

+++ Assets/GGJ2025/Scripts/UI/Buttons/DeleteSaveButton.cs
@@ -1,5 +1,6 @@
 public class DeleteSaveButton : BRButton {
     protected override void OnClick() {
+        if (UIController.isPrevented) return;
         base.OnClick();
         LevelManager.Get().Level = SaveSystem.RemoveFile();
     }
